RDKBDEV-2465: add comment that CMMGMT_CM_DHCP_INFO..
Reason for change:
cm: add comment that CMMGMT_CM_DHCP_INFO must exactly match COSA_CM_DHCP_INFO in ccsp-cm-agent

It's bad that this struct defintion is duplicated between the CM HAL
and ccsp-cm-agent. As a temporary solution until a real fix is made,
add a warning to cm_hal.h to document that the struct definition must
be kept in sync with the ccsp-cm-agent version.

Test Procedure: Sanity.
Risks: None.

Change-Id: Ie549ad5dec5c283f8cd74b59b8d93e6da4440697
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>